fix: Update CI test configurations for Gemma3 compatibility
- Change gemma3-transformers test to use model_impl='auto' instead of 'transformers'
  to avoid missing generate_attention_masks() method in TransformersMultiModalForCausalLM
- Change quantized models test to use dtype='bfloat16' instead of 'half'
  since gemma3_text models don't support float16 due to numerical instability

These are test configuration fixes for pre-existing bugs unrelated to GGUF changes.

+++ tests/models/quantization/test_gguf.py
@@ -155,7 +155,7 @@ def check_model_outputs(
     "model",
     [pytest.param(test_config, marks=test_config.marks) for test_config in MODELS],
 )
-@pytest.mark.parametrize("dtype", ["half"])
+@pytest.mark.parametrize("dtype", ["bfloat16"])
 @pytest.mark.parametrize("max_tokens", [32])
 @pytest.mark.parametrize("num_logprobs", [5])
 @pytest.mark.parametrize("tp_size", [1])
